extract navigation constant
